fix node issues and missing property in save data

+++ Assets/_Project/Scripts/VisualScripting/OperatorHandle/DivisionHandler.cs
@@ -4,7 +4,7 @@ namespace Loykas.Scripting
     {
         public DivisionHandler()
         {
-            Operator<double, double>((a, b) => a / b);
+            Operator<float, float>((a, b) => a / b);
         }
     }
 }
